Match Exceptions to the rows they came from
When importing it may be nice to be able to tell the user that a particular row is where the exception occurred. This allows us to use the loop index to map an exception to a particular row so that programs could provide additional information to the users if desired.