Optionally clear entity manager
Some bugs occur due to flushing the entity manager. This provides
a user the option to clear or not.

This seems to be left over from the old library where it had built
in batch handling. As if a developer/user would batch process in the
same memory process. This is likely not to be the case I would think.
So could be reviewed for removal. I know I've hit 3 different bugs/issues
because of calling clear().