Improve Result interface
Allow the Offset filter to stop the process function loop. The impetus is the
lack of detail the result object contains - particularly when using the Offset
filter. When processing a file in chunks the StepAggregator reports the number
of successfully imported results and errors but not the number of skipped. This
allows a user to seek the reader, and then process a batch and then stop.
Receiving an accurate set of processed,skipped and errors.