$tableName into backticks
In case a table name is a reserved word like "where" (which would be stupid ;)) or there is e.g. a "-" in the table name, say `$tableName` would be `Customer-Names` this throws multiple errors. If you put the table name in backticks, it works.

more: https://stackoverflow.com/questions/1962859/importance-of-backtick-around-table-name-in-mysql-query